Remove lambda closure value gap at 0xff
Historically the number of lambda constants are saved as the constant
no. 0xff, if the number exceeds 0xff, because in the function header
was only a byte for this number. Since LDMud 3.5.0 the function header
was redesigned and can contain the full number, so this isn't needed
anymore. Also give an error message if there are more than 0xffff
constants, because the bytecode doesn't support a larger number.