sprintf: restore original null character behaviour.
Previous to 3.3.x `sprintf("hello%cworld\n", 0);` would evaluate to
`"helloworld\n"` - the null character is treated as an empty string. In
3.3.720 however it evaluates to `"hello"` - the null character truncates
the string c-style. This is probably the more sensible behaviour but we
have a lot of legacy lib code that expected to be able to use `0` as
`""` for a `%c` character specifier and now text (like inventory
output!) is suddenly and incorrectly truncated.

Finding all of the instances of this across the lib is a tricky runtime
problem. Since our lib code never expects to be able to `sprintf` a null
character (it wasn't possible previously!) it seems much easier to patch
the driver to restore the original behaviour. Sorry not sorry.