restore --enable-use-set-light and set_light efun.
Some large and unwieldy libs (cough) aren't ready to migrate to
providing `efun::set_light()` from the lib. The diff required to restore
support is small and straight-forward. I've also elected to not provide
support for `OIB_TOTAL_LIGHT` with `object_info`. In my case there was
only ~1 consumer of that API and avoiding restoring support for it means
not needing to change `sys/objectinfo.h` in the lib.