pkg-python: hack to fix python_reload().
In `call_python_efun` from `pkg-python.c` I've noticed buggy behaviour
from the `python_reload()` python efun provided by the python-efuns
Python package. Calling the reload efun from a non-interactive object
seems to break subsequent external calls from Python->LPC. In practice
it seems the root cause is `this_object()` getting saved before the
python code is reloaded in a way that causes all subsequent external
calls to use the saved object as if it were `this_object()` for the
external call. If the saved object happens to be an interactive most
things seem to "work" but with the incorrect context. If the saved
object is not interactive, things silently break.

In the short term it seems like skipping the call to
`python_save_context()` when executing the `python_reload` efun is
enough to prevent this bug from occurring. Subsequent external calls
correctly have no associated context object and get master as
a placeholder.